+++ platform/linux/window-fbdev.h
@@ -5,12 +5,14 @@
 #error This header file is internal to the cgd platform module and is not intended to be used elsewhere
 #endif /* P_INTERNAL_GUARD__ */
 
+#include "../window.h"
 #include <core/int.h>
 #include <core/pixel.h>
 #include <core/shapes.h>
 #include <stdbool.h>
 #include <termios.h>
 #include <pthread.h>
+#include <semaphore.h>
 #include <linux/fb.h>
 
 struct window_fbdev_listener {
@@ -19,12 +21,15 @@ struct window_fbdev_listener {
     pthread_t thread;
     pthread_mutex_t buf_mutex;
 
-    struct pixel_flat_data front_buffer;
-    _Atomic bool page_flip_pending;
-
-    i32 *fd_p;
-    const rect_t *win_rect_p, *display_rect_p;
+    sem_t page_flip_pending;
+    /* The thread only needs to read from the buffer
+     * and write to the map */
+    const struct pixel_flat_data *_Atomic front_buffer_p;
     void *const *map_p;
+
+    /* All of these should be read-only for the thread */
+    const i32 *fd_p;
+    const rect_t *win_rect_p, *display_rect_p;
     const u32 *stride_p;
 };
 
@@ -37,7 +42,7 @@ struct window_fbdev {
     u64 mem_size;
 
     struct pixel_flat_data back_buffer;
-    /* The front buffer is in the listener */
+    struct pixel_flat_data front_buffer;
 
     struct window_fbdev_listener listener;
 
@@ -59,8 +64,7 @@ i32 window_fbdev_open(struct window_fbdev *win,
     const rect_t *area, const u32 flags);
 void window_fbdev_close(struct window_fbdev *win);
 
-struct pixel_flat_data * window_fbdev_swap_buffers(struct window_fbdev *win);
-void window_fbdev_render(struct window_fbdev *win,
-    const struct pixel_flat_data *pixels);
+struct pixel_flat_data * window_fbdev_swap_buffers(struct window_fbdev *win,
+    const enum p_window_present_mode present_mode);
 
 #endif /* P_WINDOW_FBDEV_H_ */