+++ platform/linux/window-x11.c
@@ -31,9 +31,6 @@
 #define P_INTERNAL_GUARD__
 #include "window-x11-events.h"
 #undef P_INTERNAL_GUARD__
-#define P_INTERNAL_GUARD__
-#include "window-internal.h"
-#undef P_INTERNAL_GUARD__
 
 #define MODULE_NAME "window-x11"
 
@@ -301,7 +298,7 @@ void window_X11_render(struct window_x11 *win, struct pixel_flat_data *fb)
 {
     u_check_params(win != NULL);
 
-    if (win->gpu_acceleration != WINDOW_ACCELERATION_NONE)
+    if (win->gpu_acceleration != P_WINDOW_ACCELERATION_NONE)
         return;
 
     s_assert(fb != NULL && fb->buf != NULL && fb->w > 0 && fb->h > 0,
@@ -383,12 +380,14 @@ void window_X11_deregister_mouse(struct window_x11 *win)
     win->registered_mouse = NULL;
 }
 
-void window_X11_set_acceleration(struct window_x11 *win, i32 val_i32)
+i32 window_X11_set_acceleration(struct window_x11 *win,
+    enum p_window_acceleration new_val)
 {
-    enum window_acceleration val = val_i32;
 
-    u_check_params(win != NULL && val >= 0 && val < WINDOW_ACCELERATION_MAX_);
-    win->gpu_acceleration = val;
+    u_check_params(win != NULL &&
+        new_val >= 0 && new_val < P_WINDOW_ACCELERATION_MAX_);
+    win->gpu_acceleration = new_val;
+    return 0;
 }
 
 static i32 init_fb(struct window_x11 *win, u32 fb_w, u32 fb_h)