+++ render/renderer-main.c
@@ -7,6 +7,6 @@ void renderer_main(void *arg)
 {
     struct r_ctx_thread_info *info = arg;
     p_mt_mutex_lock(&info->mutex);
-    p_mt_cond_wait(info->running, info->mutex);
+    p_mt_cond_wait(info->cond, info->mutex);
     p_mt_thread_exit();
 }